% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_rate.R
\name{CalcEigenStationary}
\alias{CalcEigenStationary}
\title{Calculate the Stationary Distribution of a First-Order Markov Chain from its
Transition Matrix}
\usage{
CalcEigenStationary(trans_mat)
}
\arguments{
\item{trans_mat}{Transition matrix}
}
\value{
vector representing the stationary distribution of the Markov chain
}
\description{
Calculate the Stationary Distribution of a First-Order Markov Chain from its
Transition Matrix
}
\examples{
tm1 <- matrix(c(0,0,1,1,0,0,0,1,0), 3,3, T)
sm1 <- CalcEigenStationary(tm1)
tm2 <- matrix(c(.3, 0.7, 0.8, 0.2), 2,2, T)
sm2 <- CalcEigenStationary(tm2)
}

