% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_rate.R
\name{CalcEmpiricalStationary}
\alias{CalcEmpiricalStationary}
\title{Calculation of the Empirical Stationary Distribution from an observed Markov Chain (first-order)}
\usage{
CalcEmpiricalStationary(event_seq, state_space)
}
\arguments{
\item{event_seq}{Observed sequence of events as a Markov chain}

\item{state_space}{State space of the observed Markov chain}
}
\value{
Estimate of the stationary distribution
}
\description{
Calculation of the Empirical Stationary Distribution from an observed Markov Chain (first-order)
}
\examples{
t_mat <- matrix(c(0.3, 0.7, 0.6, 0.4), 2,2, T)
sim_mc <- SimulateMarkovChain(t_mat, n_sims = 500)
true_sm <- CalcEigenStationary(t_mat)
est_sm <- CalcEmpiricalStationary(sim_mc, 1:2)
}

