% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_rate.R
\name{CalcEmpStat_Mth_Order}
\alias{CalcEmpStat_Mth_Order}
\title{Calculation of the Empirical Stationary Distribution from an observed Markov Chain (Arbitrary-Order)}
\usage{
CalcEmpStat_Mth_Order(event_seq, state_space, mc_order = 1)
}
\arguments{
\item{event_seq}{Observed sequence of events as a Markov chain}

\item{state_space}{State space of the observed Markov chain}

\item{mc_order}{Order of the Markov Chain}
}
\value{
Estimate of the stationary distribution
}
\description{
Calculation of the Empirical Stationary Distribution from an observed Markov Chain (Arbitrary-Order)
}
\examples{
t_mat <- matrix(c(0.3, 0.7, 0.6, 0.4), 2,2, T)
sim_mc <- SimulateMarkovChain(t_mat, n_sims = 10000)
tc2 <- CalcTC_Mth_Order(sim_mc, 1:2, 2)
tm2 <- CalcTransitionMatrix(tc2)
eig_sm <- CalcEigenStationary(tm2)
emp_sm <- CalcEmpStat_Mth_Order(sim_mc, 1:2, 2)
print(eig_sm)
print(emp_sm)
}

