% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_rate.R
\name{CalcEntropyRate}
\alias{CalcEntropyRate}
\title{Calculate the Entropy Rate of a Finite Markov Chain}
\usage{
CalcEntropyRate(event_seq, state_space, mc_order = 1,
  stat_method = "Empirical")
}
\arguments{
\item{event_seq}{Observed sequence of events}

\item{state_space}{State space of the observed process}

\item{mc_order}{Order of the Markov Chain}

\item{stat_method}{Method for computing the Stationary distribution}
}
\value{
Estimate of the Entropy Rate
}
\description{
Calculate the Entropy Rate of a Finite Markov Chain
}
\examples{
t_mat <- matrix(c(0.3, 0.7, 0.6, 0.4), 2,2, T)
sim_mc <- SimulateMarkovChain(t_mat, n_sims = 50000)
CalcMarkovEntropyRate(t_mat, CalcEigenStationary(t_mat))
CalcEntropyRate(sim_mc, 1:2, mc_order = 1, stat_method="Empirical")
CalcEntropyRate(sim_mc, 1:2, mc_order = 1, stat_method="Eigen")
}

