% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_rate.R
\name{CalcTC_Mth_Order}
\alias{CalcTC_Mth_Order}
\title{Compute a Matrix of Transition Counts from an Observed Sequence (Arbitrary Order)}
\usage{
CalcTC_Mth_Order(event_seq, state_space, mc_order = 1)
}
\arguments{
\item{event_seq}{Vector of observations.}

\item{state_space}{vector composed of all of the state space}

\item{mc_order}{order of the Markov chain.  Defaults to 1.}
}
\value{
Returns a matrix of transition counts of a given order
}
\description{
Compute a Matrix of Transition Counts from an Observed Sequence (Arbitrary Order)
}
\examples{
t_mat <- matrix(c(0.3, 0.7, 0.6, 0.4), 2,2, T)
sim_mc <- SimulateMarkovChain(t_mat, n_sims = 500)
tc2 <- CalcTC_Mth_Order(sim_mc, 1:2, mc_order=2)
}

