% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_rate.R
\name{CalcTransitionCounts}
\alias{CalcTransitionCounts}
\title{Compute a Matrix of Transition Counts from an Observed Sequence}
\usage{
CalcTransitionCounts(event_seq, n_states = length(unique(event_seq)))
}
\arguments{
\item{event_seq}{Vector of observations.}

\item{n_states}{Total number of expected unique states.}
}
\value{
Matrix representing the number of transitions from each state to all other states.
}
\description{
Compute a Matrix of Transition Counts from an Observed Sequence
}
\examples{
t_mat <- matrix(c(0.3, 0.7, 0.6, 0.4), 2,2, T)
sim_mc <- SimulateMarkovChain(t_mat, n_sims = 500)
tc <- CalcTransitionCounts(sim_mc)
}
