% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccber.R
\docType{package}
\name{ccber}
\alias{ccber}
\alias{ccber-package}
\title{ccber: A package for estimating behavioral entropy rate from observational
data}
\description{
The ccber package has two categories of functions:
estimating entropy rate, analyzing files
}
\section{Estimating Entropy Rate}{

These are the functions which estimate entropy rate...
}

\section{Analzying files}{

These help analyze files
}

