% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_rate.R
\name{SimulateMarkovChain}
\alias{SimulateMarkovChain}
\title{Simulate a Markov Chain Given a Transition Matrix}
\usage{
SimulateMarkovChain(trans_mat, n_sims = 100)
}
\arguments{
\item{trans_mat}{A row-stochastic matrix representing a Markov transition matrix}

\item{n_sims}{Number of observations for simulation length. Default of 100.}
}
\value{
A vector of observations from a Markov chain
}
\description{
Simulate a Markov Chain Given a Transition Matrix
}
\examples{
t_mat <- matrix(c(0.3, 0.7, 0.6, 0.4), 2,2, T)
sim_mc <- SimulateMarkovChain(t_mat)
sim_mc <- SimulateMarkovChain(t_mat, n_sims = 500)
}

